+++ src/Option.ts
@@ -1216,27 +1216,6 @@ export function exists<A>(predicate: Predicate<A>): (ma: Option<A>) => boolean {
   return (ma) => (isNone(ma) ? false : predicate(ma.value))
 }
 
-/**
- * Returns a `Refinement` (i.e. a custom type guard) from a `Option` returning function.
- * This function ensures that a custom type guard definition is type-safe.
- *
- * ```ts
- * import { some, none, getRefinement } from 'fp-ts/Option'
- *
- * type A = { type: 'A' }
- * type B = { type: 'B' }
- * type C = A | B
- *
- * const isA = (c: C): c is A => c.type === 'B' // <= typo but typescript doesn't complain
- * const isA = getRefinement<C, A>(c => (c.type === 'B' ? some(c) : none)) // static error: Type '"B"' is not assignable to type '"A"'
- * ```
- *
- * @since 2.0.0
- */
-export function getRefinement<A, B extends A>(getOption: (a: A) => Option<B>): Refinement<A, B> {
-  return (a: A): a is B => isSome(getOption(a))
-}
-
 // -------------------------------------------------------------------------------------
 // do notation
 // -------------------------------------------------------------------------------------
@@ -1316,6 +1295,16 @@ export const sequenceArray: <A>(arr: ReadonlyArray<Option<A>>) => Option<Readonl
 // -------------------------------------------------------------------------------------
 // deprecated
 // -------------------------------------------------------------------------------------
+
+/**
+ * Use `Refinement` module instead.
+ *
+ * @since 2.0.0
+ * @deprecated
+ */
+export function getRefinement<A, B extends A>(getOption: (a: A) => Option<B>): Refinement<A, B> {
+  return (a: A): a is B => isSome(getOption(a))
+}
 
 /**
  * @category combinators